reconstitute directories #1
