--- todo/java/TodoForMybatis.java
@@ -1,50 +1,0 @@
-package todo.domain.model;
-
-import java.io.Serializable;
-import java.util.Date;
-
-public class Todo implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private String todoId;
-
-    private String todoTitle;
-
-    private boolean finished;
-
-    private Date createdAt;
-
-    public String getTodoId() {
-        return todoId;
-    }
-
-    public void setTodoId(String todoId) {
-        this.todoId = todoId;
-    }
-
-    public String getTodoTitle() {
-        return todoTitle;
-    }
-
-    public void setTodoTitle(String todoTitle) {
-        this.todoTitle = todoTitle;
-    }
-
-    public boolean isFinished() {
-        return finished;
-    }
-
-    public void setFinished(boolean finished) {
-        this.finished = finished;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-}
-
