--- todo/java/TodoService.java
@@ -1,16 +1,0 @@
-package todo.domain.service.todo;
-
-import java.util.Collection;
-
-import todo.domain.model.Todo;
-
-public interface TodoService {
-    Collection<Todo> findAll();
-
-    Todo create(Todo todo);
-
-    Todo finish(String todoId);
-
-    void delete(String todoId);
-}
-
