+++ todo/scripts/generate-security-project.sh
@@ -51,35 +51,35 @@ mkdir -p ./first-spring-security-mybatis3/src/main/webapp/WEB-INF/views/account/
 mkdir -p ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-web/src/main/webapp/WEB-INF/views/account/
 
 # cp security sources
-cp -p ./todo/java/Account.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/model
-cp -p ./todo/java/Account.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/model
+cp -p ./todo/src/main/java/security/com/example/security/domain/model/Account.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/model
+cp -p ./todo/src/main/java/security/com/example/security/domain/model/Account.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/model
 
-cp -p ./todo/java/AccountRepository.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/repository/account/
-cp -p ./todo/java/AccountRepository.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/repository/account/
+cp -p ./todo/src/main/java/security/com/example/security/domain/repository/account/AccountRepository.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/repository/account/
+cp -p ./todo/src/main/java/security/com/example/security/domain/repository/account/AccountRepository.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/repository/account/
 
-cp -p ./todo/xml/AccountRepository.xml ./first-spring-security-mybatis3/src/main/resources/com/example/security/domain/repository/account/
-cp -p ./todo/xml/AccountRepository.xml ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/resources/com/example/security/domain/repository/account/
+cp -p ./todo/src/main/resources/com/example/security/domain/repository/account/AccountRepository.xml ./first-spring-security-mybatis3/src/main/resources/com/example/security/domain/repository/account/
+cp -p ./todo/src/main/resources/com/example/security/domain/repository/account/AccountRepository.xml ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/resources/com/example/security/domain/repository/account/
 
-cp -p ./todo/java/AccountSharedService.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/account/
-cp -p ./todo/java/AccountSharedService.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/account/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/account/AccountSharedService.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/account/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/account/AccountSharedService.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/account/
 
-cp -p ./todo/java/AccountSharedServiceImpl.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/account/
-cp -p ./todo/java/AccountSharedServiceImpl.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/account/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/account/AccountSharedServiceImpl.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/account/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/account/AccountSharedServiceImpl.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/account/
 
-cp -p ./todo/java/AccountController.java ./first-spring-security-mybatis3/src/main/java/com/example/security/app/account/
-cp -p ./todo/java/AccountController.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-web/src/main/java/com/example/security/app/account/
+cp -p ./todo/src/main/java/security/com/example/security/app/account/AccountController.java ./first-spring-security-mybatis3/src/main/java/com/example/security/app/account/
+cp -p ./todo/src/main/java/security/com/example/security/app/account/AccountController.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-web/src/main/java/com/example/security/app/account/
 
-cp -p ./todo/java/SampleUserDetails.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/userdetails/
-cp -p ./todo/java/SampleUserDetails.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/userdetails/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/userdetails/SampleUserDetails.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/userdetails/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/userdetails/SampleUserDetails.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/userdetails/
 
-cp -p ./todo/java/SampleUserDetailsService.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/userdetails/
-cp -p ./todo/java/SampleUserDetailsService.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/userdetails/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/userdetails/SampleUserDetailsService.java ./first-spring-security-mybatis3/src/main/java/com/example/security/domain/service/userdetails/
+cp -p ./todo/src/main/java/security/com/example/security/domain/service/userdetails/SampleUserDetailsService.java ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-domain/src/main/java/com/example/security/domain/service/userdetails/
 
-cp -p ./todo/sql/H2-*.sql ./first-spring-security-mybatis3/src/main/resources/database/
-cp -p ./todo/sql/H2-*.sql ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-env/src/main/resources/database/
+cp -p ./todo/src/main/resources/database/H2-*.sql ./first-spring-security-mybatis3/src/main/resources/database/
+cp -p ./todo/src/main/resources/database/H2-*.sql ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-env/src/main/resources/database/
 
-cp -p ./todo/jsp/login.jsp ./first-spring-security-mybatis3/src/main/webapp/
-cp -p ./todo/jsp/login.jsp ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-web/src/main/webapp/
+cp -p ./todo/src/main/webapp/WEB-INF/login.jsp ./first-spring-security-mybatis3/src/main/webapp/
+cp -p ./todo/src/main/webapp/WEB-INF/login.jsp ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-web/src/main/webapp/
 
-cp -p ./todo/jsp/view.jsp ./first-spring-security-mybatis3/src/main/webapp/WEB-INF/views/account/
-cp -p ./todo/jsp/view.jsp ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-web/src/main/webapp/WEB-INF/views/account/
+cp -p ./todo/src/main/webapp/WEB-INF/views/account/view.jsp ./first-spring-security-mybatis3/src/main/webapp/WEB-INF/views/account/
+cp -p ./todo/src/main/webapp/WEB-INF/views/account/view.jsp ./first-spring-security-mybatis3-multi/first-spring-security-mybatis3-multi-web/src/main/webapp/WEB-INF/views/account/
