+++ todo/src/main/java/jpa/todo/domain/model/Todo.java
@@ -1,0 +1,61 @@
+package todo.domain.model;
+
+import java.io.Serializable;
+import java.util.Date;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+
+// (1)
+@Entity
+@Table(name = "todo")
+public class Todo implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    // (2)
+    @Id
+    private String todoId;
+
+    private String todoTitle;
+
+    private boolean finished;
+
+    // (3)
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date createdAt;
+
+    public String getTodoId() {
+        return todoId;
+    }
+
+    public void setTodoId(String todoId) {
+        this.todoId = todoId;
+    }
+
+    public String getTodoTitle() {
+        return todoTitle;
+    }
+
+    public void setTodoTitle(String todoTitle) {
+        this.todoTitle = todoTitle;
+    }
+
+    public boolean isFinished() {
+        return finished;
+    }
+
+    public void setFinished(boolean finished) {
+        this.finished = finished;
+    }
+
+    public Date getCreatedAt() {
+        return createdAt;
+    }
+
+    public void setCreatedAt(Date createdAt) {
+        this.createdAt = createdAt;
+    }
+}
