+++ todo/src/main/java/mybatis3/todo/domain/service/todo/TodoServiceImpl.java
@@ -1,0 +1,100 @@
+package todo.domain.service.todo;
+
+import java.util.Collection;
+import java.util.Date;
+import java.util.UUID;
+
+import javax.inject.Inject;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.terasoluna.gfw.common.exception.BusinessException;
+import org.terasoluna.gfw.common.exception.ResourceNotFoundException;
+import org.terasoluna.gfw.common.message.ResultMessage;
+import org.terasoluna.gfw.common.message.ResultMessages;
+
+import todo.domain.model.Todo;
+import todo.domain.repository.todo.TodoRepository;
+
+@Service// (1)
+@Transactional // (2)
+public class TodoServiceImpl implements TodoService {
+
+    private static final long MAX_UNFINISHED_COUNT = 5;
+
+    @Inject// (3)
+    TodoRepository todoRepository;
+
+    // (4)
+    public Todo findOne(String todoId) {
+        Todo todo = todoRepository.findOne(todoId);
+        if (todo == null) {
+            // (5)
+            ResultMessages messages = ResultMessages.error();
+            messages.add(ResultMessage
+                    .fromText("[E404] The requested Todo is not found. (id="
+                            + todoId + ")"));
+            // (6)
+            throw new ResourceNotFoundException(messages);
+        }
+        return todo;
+    }
+
+    @Override
+    @Transactional(readOnly = true) // (7)
+    public Collection<Todo> findAll() {
+        return todoRepository.findAll();
+    }
+
+    @Override
+    public Todo create(Todo todo) {
+        long unfinishedCount = todoRepository.countByFinished(false);
+        if (unfinishedCount >= MAX_UNFINISHED_COUNT) {
+            ResultMessages messages = ResultMessages.error();
+            messages.add(ResultMessage
+                    .fromText("[E001] The count of un-finished Todo must not be over "
+                            + MAX_UNFINISHED_COUNT + "."));
+            // (8)
+            throw new BusinessException(messages);
+        }
+
+        // (9)
+        String todoId = UUID.randomUUID().toString();
+        Date createdAt = new Date();
+
+        todo.setTodoId(todoId);
+        todo.setCreatedAt(createdAt);
+        todo.setFinished(false);
+
+        todoRepository.create(todo);
+        /* REMOVE THIS LINE IF YOU USE JPA
+            todoRepository.save(todo); // 10
+           REMOVE THIS LINE IF YOU USE JPA */
+
+        return todo;
+    }
+
+    @Override
+    public Todo finish(String todoId) {
+        Todo todo = findOne(todoId);
+        if (todo.isFinished()) {
+            ResultMessages messages = ResultMessages.error();
+            messages.add(ResultMessage
+                    .fromText("[E002] The requested Todo is already finished. (id="
+                            + todoId + ")"));
+            throw new BusinessException(messages);
+        }
+        todo.setFinished(true);
+        todoRepository.update(todo);
+        /* REMOVE THIS LINE IF YOU USE JPA
+            todoRepository.save(todo); // (11)
+           REMOVE THIS LINE IF YOU USE JPA */
+        return todo;
+    }
+
+    @Override
+    public void delete(String todoId) {
+        Todo todo = findOne(todoId);
+        todoRepository.delete(todo);
+    }
+}
