+++ todo/src/main/java/security/com/example/security/domain/service/account/AccountSharedServiceImpl.java
@@ -1,0 +1,30 @@
+package com.example.security.domain.service.account;
+
+import javax.inject.Inject;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.terasoluna.gfw.common.exception.ResourceNotFoundException;
+
+import com.example.security.domain.model.Account;
+import com.example.security.domain.repository.account.AccountRepository;
+
+@Service
+public class AccountSharedServiceImpl implements AccountSharedService {
+    @Inject
+    AccountRepository accountRepository;
+
+    @Transactional(readOnly=true)
+    @Override
+    public Account findOne(String username) {
+        // (1)
+        Account account = accountRepository.findOne(username);
+        // (2)
+        if (account == null) {
+            throw new ResourceNotFoundException("The given account is not found! username="
+                    + username);
+        }
+        return account;
+    }
+
+}
