+++ todo/src/main/java/todo/domain/service/todo/TodoService.java
@@ -1,0 +1,15 @@
+package todo.domain.service.todo;
+
+import java.util.Collection;
+
+import todo.domain.model.Todo;
+
+public interface TodoService {
+    Collection<Todo> findAll();
+
+    Todo create(Todo todo);
+
+    Todo finish(String todoId);
+
+    void delete(String todoId);
+}
